split useOptions into new file
