+++ packages/grafana-ui/src/components/Combobox/MultiCombobox.tsx
@@ -18,13 +18,14 @@ import { NotFoundError } from './MessageRows';
 import { OptionListItem } from './OptionListItem';
 import { SuffixIcon } from './SuffixIcon';
 import { ValuePill } from './ValuePill';
-import { itemFilter, itemToString } from './filter';
+import { itemToString } from './filter';
 import { getComboboxStyles, MENU_OPTION_HEIGHT, MENU_OPTION_HEIGHT_DESCRIPTION } from './getComboboxStyles';
 import { getMultiComboboxStyles } from './getMultiComboboxStyles';
 import { ALL_OPTION_VALUE, ComboboxOption } from './types';
 import { useComboboxFloat } from './useComboboxFloat';
 import { MAX_SHOWN_ITEMS, useMeasureMulti } from './useMeasureMulti';
 import { useMultiInputAutoSize } from './useMultiInputAutoSize';
+import { useOptions } from './useOptions';
 
 interface MultiComboboxBaseProps<T extends string | number> extends Omit<ComboboxBaseProps<T>, 'value' | 'onChange'> {
   value?: T[] | Array<ComboboxOption<T>>;
@@ -34,71 +35,9 @@ interface MultiComboboxBaseProps<T extends string | number> extends Omit<Combobo
 
 export type MultiComboboxProps<T extends string | number> = MultiComboboxBaseProps<T> & AutoSizeConditionals;
 
-// A hook to handle sync or async options
-// intended to be used by both MultiCombobox and Combobox
-function useOptions<T extends string | number>(options: MultiComboboxBaseProps<T>['options']) {
-  const isAsync = typeof options === 'function';
-  const [asyncOptions, setAsyncOptions] = useState<Array<ComboboxOption<T>>>([]);
-  const [asyncLoading, setAsyncLoading] = useState(false);
-  const [asyncError, setAsyncError] = useState(false);
-
-  // This hook keeps its own inputValue state (rather than accepting it as an arg) because it needs to be
-  // told it for async options loading anyway.
-  const [userTypedSearch, setUserTypedSearch] = useState('');
-
-  const loadOptionsWhenUserTypes = useCallback(
-    (inputValue: string) => {
-      console.log('loadOptionsWhenUserTypes', inputValue);
-      if (!isAsync) {
-        setUserTypedSearch(inputValue);
-        return;
-      }
-
-      setAsyncLoading(true);
-
-      options(inputValue)
-        .then((options) => {
-          setAsyncOptions(options);
-          setAsyncLoading(false);
-          setAsyncError(false);
-        })
-        .catch((error) => {
-          setAsyncError(true);
-          setAsyncLoading(false);
-
-          if (error) {
-            console.error('Error loading async options for Combobox', error);
-          }
-        });
-    },
-    [options, isAsync]
-  );
-
-  const finalOptions = useMemo(() => {
-    if (isAsync) {
-      return asyncOptions;
-    } else {
-      return options.filter(itemFilter(userTypedSearch));
-    }
-  }, [options, asyncOptions, isAsync, userTypedSearch]);
-
-  return { options: finalOptions, loadOptionsWhenUserTypes, asyncLoading, asyncError };
-}
-
 export const MultiCombobox = <T extends string | number>(props: MultiComboboxProps<T>) => {
-  const {
-    options,
-    placeholder,
-    onChange,
-    value,
-    width,
-    enableAllOption,
-    invalid,
-    loading,
-    disabled,
-    minWidth,
-    maxWidth,
-  } = props;
+  const { placeholder, onChange, value, width, enableAllOption, invalid, loading, disabled, minWidth, maxWidth } =
+    props;
 
   const { options: newOptions, loadOptionsWhenUserTypes /*, asyncLoading, asyncError */ } = useOptions(props.options);
